requirements.txt
discord.py

discord-webhook
pixivpy
python-twitter
